Remove KMF support and add missing ocs provider's files.
This commit removes the Key Management Framework (KMF) support from the project. The changes
include the deletion of KMF specific files and the removal of KMF related configuration options
and flags across various Makefile.in files. Additionally, the missing files for the initial
implementation of the providers were added.

Key changes:
- Deletion of src/drivers/kmf directory and its contents, which were specific to KMF support.
- Removal of KMF related configuration options (--enable-kmf) and variables (kmf_cflags, kmf_ldadd,
  kmf_libflags, kmf_prefix) from configure, configure.ac, and various Makefile.in files.
- Added missing OCS provider files.